[Tizen.Applications.Cion] Fix namespace
Fix namespace to Tizen.Applications.Cion from Tizen.Applications.
