+++ src/Tizen.Applications.Cion/Tizen.Applications.Cion/CionErrorFactory.cs
@@ -16,7 +16,7 @@
 
 using System;
 
-namespace Tizen.Applications
+namespace Tizen.Applications.Cion
 {
     internal static class CionErrorFactory
     {
