+++ src/Tizen.Applications.Cion/Tizen.Applications.Cion/ConnectionStatus.cs
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-namespace Tizen.Applications
+namespace Tizen.Applications.Cion
 {
     /// <summary>
     /// Enumeration for connection status.
