+++ src/Tizen.Applications.Cion/Tizen.Applications.Cion/GroupSafeHandle.cs
@@ -18,20 +18,15 @@ using System;
 using System.ComponentModel;
 using System.Runtime.InteropServices;
 
-namespace Tizen.Applications
+namespace Tizen.Applications.Cion
 {
-    internal sealed class PayloadSafeHandle : SafeHandle
+    internal sealed class GroupSafeHandle : SafeHandle
     {
         [EditorBrowsable(EditorBrowsableState.Never)]
-        public PayloadSafeHandle() : base(IntPtr.Zero, true)
+        public GroupSafeHandle() : base(IntPtr.Zero, true)
         {
         }
-        
-        internal PayloadSafeHandle(IntPtr existingHandle, bool ownsHandle) : base(IntPtr.Zero, ownsHandle)
-        {
-            SetHandle(existingHandle);
-        }
-        
+
         [EditorBrowsable(EditorBrowsableState.Never)]
         public override bool IsInvalid
         {
@@ -40,7 +35,7 @@ namespace Tizen.Applications
 
         protected override bool ReleaseHandle()
         {
-            Interop.CionPayload.CionPayloadDestroy(this.handle);
+            Interop.CionGroup.CionGroupDestroy(this.handle);
             SetHandle(IntPtr.Zero);
             return true;
         }
