+++ src/Tizen.Applications.Cion/Tizen.Applications.Cion/PayloadAsyncResult.cs
@@ -16,7 +16,7 @@
 
 using System;
 
-namespace Tizen.Applications
+namespace Tizen.Applications.Cion
 {
     /// <summary>
     /// A class to represent result of payload.
