+++ src/Tizen.Applications.Cion/Tizen.Applications.Cion/PeerInfo.cs
@@ -16,7 +16,7 @@
 
 using System;
 
-namespace Tizen.Applications
+namespace Tizen.Applications.Cion
 {
     /// <summary>
     /// A class to represent peer info.
