+++ src/Tizen.Applications.Cion/Tizen.Applications.Cion/SecurityInfo.cs
