+++ src/Tizen.Applications.Cion/Tizen.Applications.Cion/SecuritySafeHandle.cs
